github-push-issues.py: Create milestones via the GitHub API
Endpoint docs at [1] and content-type docs at [2].  Most of the urllib
infrastructure will be shared with Issue, so the guts are in
_Entry.create() with helpers for the issue- and milestone-specific
bits.

The commented-out Request(..., method='POST') is because Python 2
doesn't support the argument [3], but once we're Python-3-only we
should use it [4] (because explicit is better than implicit).

[1]: https://developer.github.com/v3/issues/milestones/#create-a-milestone
[2]: https://developer.github.com/v3/media/
[3]: https://docs.python.org/2/library/urllib2.html#urllib2.Request
[4]: https://docs.python.org/3/library/urllib.request.html#urllib.request.Request