github-push-issues.py: Rework walk() returns to something zip-compatible
We won't be able to use open() directly on listings returned from
ZipFile, etc., so adjust this to return a per-directory dict of
openers with filename keywords.

The functools.partial call locks in the local 'path' variable to avoid
having it mutate out from under us as we iterate over filenames
building openers.